MEDICINAL FLOWER DETECTION USING CNN ALGORITHM
In the field of Ayurveda, the accurate identification of medicinal flowers is a critical aspect of ensuring the efficiency and safety of
traditional remedies.Traditionally, medical professionals have relied on their visual acumen and olfactory senses to identify these therapeutic plants. However, the potential for human error in this process necessitates a more precise and reliable method. This approach employs a deep learning technique namely Convolutional Neural Networks (CNN), to enhance the accuracy of medicinal flower identification. Theproject aims to develop a robust model capable of classifying medicinal flowers with a high degree of precision, reducing the risks associated with mis-identification. The methodology involves the use of a diverse and comprehensive dataset comprising photographs of specific therapeutic flowers. Dandelion, daisy, and orchid are exemplified as notable candidates due to their frequent application in the treatment of respiratory ailments, digestive issues, skin problems, and various other health concerns in Ayurveda. The chosen Convolutional Neural Network architecture is well-suited for image classification tasks, offering the capacity to learn intricate patterns and features essential for distinguishing between different medicinal flowers. The training process involves exposing the model to a variety of images, allowing it to learn the unique characteristics associated with each flower type. The model achieved an accuracy of 94%,demonstrating its effectiveness in enhancing medicinal flower identification using CNNs.